requirements.txt
python-telegram-bot==21.6
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
